consolidate TPC tracker to one file
